LTLogger: Fix warning.
XCode 9 warns on passing an enum to vs_start, so a cast was added to
solve the warning.

+++ LTKit/LTKit/LTLogger.mm
@@ -191,7 +191,7 @@ static NSString *stringFromNSDecimalWithCurrentLocale(NSDecimal value) {
     }
 
     va_list args;
-    va_start(args, logLevel);
+    va_start(args, (NSUInteger)logLevel);
 
     NSString *logString = [[NSString alloc] initWithFormat:format arguments:args];
     for (id<LTLoggerTarget> target in targets) {
